feat: GMM synth intensity realism, soft PV mask downsampling, plainconv_ts encoder, full training resume
- synth_gmm_maisi: opt-in cross-slot intensity correlation (mu_group_ids ct/merged),
  intra-cohort member variance (sd_between_ratio), and partial-volume paint resampling
  matching the real-CT path; CPU + GPU paths in lockstep; new calibration/analysis scripts.
- mask_downsample=soft + mask_interp=bilinear: keep the area-pool fraction in [0,1] end to
  end (train only; eval stays hard). Includes an _area_pool_3d identity/integer fast path.
- arch.encoder=plainconv_ts: plans-free from-scratch PlainConvUNet encoder, e2-geometry
  default, configurable width.
- train.checkpoint=<path> now restores optimizer/scheduler/epoch/RNG so the LR schedule
  continues; train.resume_weights_only=true keeps the old behaviour.

+++ src/providers/totalseg.py
@@ -44,7 +44,8 @@ def crop_and_place(image_np, label_np, class_idx, center, T, *,
     image_t = place_image(crop_ct, out_sizes, pad_lo, T, antialias=antialias)
     lbl_small = resample_binary(crop_lbl == class_idx, tuple(out_sizes),
                                 mode=mask_downsample, occ_thr=occ_thr)
-    label_t = place_label(lbl_small, out_sizes, pad_lo, T).long()
+    # dtype preserved: long for occupancy/nearest, float32 (partial-volume fraction) for "soft"
+    label_t = place_label(lbl_small, out_sizes, pad_lo, T)
     return image_t, label_t, geom
 
 
@@ -86,7 +87,8 @@ def crop_and_place_cached(img_cache_np, label_np, class_idx, center, T, *,
     image_t = place_image(img_crop, out_sizes, pad_lo, T)  # cache->out ~identity when pitches match
     lbl_small = resample_binary(crop_lbl == class_idx, tuple(out_sizes),
                                 mode=mask_downsample, occ_thr=occ_thr)
-    label_t = place_label(lbl_small, out_sizes, pad_lo, T).long()
+    # dtype preserved: long for occupancy/nearest, float32 (partial-volume fraction) for "soft"
+    label_t = place_label(lbl_small, out_sizes, pad_lo, T)
     return image_t, label_t, geom
